+++ src/pretalx/agenda/views/talk.py
@@ -73,8 +73,6 @@ class TalkView(PermissionRequired, TemplateView):
     def get(self, request, *args, **kwargs):
         response = super().get(request, *args, **kwargs)
         csp_update = {}
-        if request.event.feature_flags["use_gravatar"]:
-            csp_update["img-src"] = "https://www.gravatar.com"
         if self.recording.get("csp_header"):
             csp_update["frame-src"] = self.recording.get("csp_header")
         response._csp_update = csp_update